Projeto biblioteca parte I

+++ IFitness/src/main/java/br/edu/ifspcjo/ads/web2/ifitness/servlets/ActivityRegisterServlet.java
@@ -18,6 +18,8 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
+import jakarta.servlet.http.HttpSession;
+
 @WebServlet("/activityRegister")
 public class ActivityRegisterServlet extends HttpServlet {
 
@@ -35,36 +37,34 @@ public class ActivityRegisterServlet extends HttpServlet {
 		Double distance = Double.parseDouble(req.getParameter("distance"));
 		Integer duration = Integer.parseInt(req.getParameter("duration"));
 
-		// buscar User logado
-		Optional<User> optional = getLoggedUser(req);
-
-		RequestDispatcher dispatcher = null;
-		
-		if(optional.isPresent()) {
+		String url;
+		HttpSession session = req.getSession(false);
+		if(session.getAttribute("user") == null) {
+			url = "/login.jsp";
+		}
+		else {
+			User user = (User)session.getAttribute("user");
 			ActivityDao activityDao = new ActivityDao(DataSourceSearcher.getInstance().getDataSource());
 			Activity activity = new Activity();
 			activity.setType(type);
 			activity.setDate(date);
 			activity.setDistance(distance);
 			activity.setDuration(duration);
-			activity.setUser(optional.get());
+			activity.setUser(user);
 			if(id == 0) {
 				if(activityDao.save(activity)) {
 					req.setAttribute("result", "registered");
-					dispatcher = req.getRequestDispatcher("/activity-register.jsp");
 				}
 			}else {
 				activity.setId(id);
 				if(activityDao.update(activity)) {
 					req.setAttribute("result", "registered");
-					dispatcher = req.getRequestDispatcher("/activity-register.jsp");
 				}
-			}	
-		} else {
-			req.setAttribute("result", "notRegistered");
-			dispatcher = req.getRequestDispatcher("/activity-register.jsp");
+			}
+			url = "/activity-register.jsp";
 		}
-
+ 
+		RequestDispatcher dispatcher = req.getRequestDispatcher(url);
 		dispatcher.forward(req, resp);
 	}
 	
@@ -73,36 +73,29 @@ public class ActivityRegisterServlet extends HttpServlet {
 		String action = req.getParameter("action");
 		Long activityId = Long.parseLong(req.getParameter("activity-id"));
 		
-		ActivityDao activityDao = new ActivityDao(DataSourceSearcher.getInstance().getDataSource());
-		Activity activity = activityDao.getActivitiesById(activityId);
-		RequestDispatcher dispatcher = null;
-		if(activity != null) {
-			if(action.equals("update")) {
-				req.setAttribute("activity", activity);
-				dispatcher = req.getRequestDispatcher("/activity-register.jsp");
-			}
-			if(action.equals("delete")) {
-				activityDao.delete(activity);
-				dispatcher = req.getRequestDispatcher("/homeServlet");
-			}
-		}else {
-			dispatcher = req.getRequestDispatcher("/homeServlet");
+		String url = null;
+		HttpSession session = req.getSession(false);
+		if(session == null || session.getAttribute("user") == null) {
+			url = "/login.jsp";
 		}
-		dispatcher.forward(req, resp);
-	}
-
-	private Optional<User> getLoggedUser(HttpServletRequest req) {
-		Optional<User> optional = Optional.empty();
-		Cookie[] cookies = req.getCookies();
-		if (cookies != null) {
-			for (Cookie c : cookies) {
-				if (c.getName().equals("loggedUser")) {
-					UserDao userDao = new UserDao(DataSourceSearcher.getInstance().getDataSource());
-					optional = userDao.getUserByEmail(c.getValue());
+		else {
+			ActivityDao activityDao = new ActivityDao(DataSourceSearcher.getInstance().getDataSource());
+			Activity activity = activityDao.getActivitiesById(activityId); 
+			if(activity != null) {
+				if(action.equals("update")) {
+					req.setAttribute("activity", activity);
+					url = "/activity-register.jsp";
 				}
+				if(action.equals("delete")) {
+					activityDao.delete(activity);
+					url = "/homeServlet";
+				}
+			}else {
+				url = "/homeServlet";
 			}
 		}
-		return optional;
+		RequestDispatcher dispatcher = req.getRequestDispatcher(url);
+		dispatcher.forward(req, resp);
 	}
 
 }
